Move treebuilder.ts to scripts/splash
