+++ scripts/splash/treebuilder.ts
